fix: Return correct difference in ft_memcmp
Memcmp was returning the total difference for
all n addresses at each pointer. It now returns
the difference between the first diverging
addresses.

Move includes to the library header.

Add an explicit makefile rule for building
object files. It is not necessary but it's
clearer and the output is better formatted this
way.